refactor: split touched files under 200 lines (decompose + reuse)
Per request, decompose the files I touched this session so each stays ≤200
lines, extracting cohesive helpers/components and removing duplication:
 - pushTextLabel → textMeasure (wrap/fit/truncate helpers)
 - buildGroupChildren → fabricImage (image paint resolution + loader)
 - ColorPickerPopover → ColorPopoverPanel + colorSwatch (shared CHECKER/PRESETS)
 - TableStyleSections → TableCellStyleFields (shared header/row cell editor)
No behaviour change.
