+++ packages/ui/src/components/RightPanel/TableStyleSections.tsx
@@ -2,8 +2,7 @@ import type { CellStyle, TableBorderStyle, TableField, TextAlign } from '@templa
 import { useTemplateStore } from '../../store/templateStore.js'
 import { NumberInput } from '../NumberInput.js'
 import { NullableColorInput } from '../NullableColorInput.js'
-import { ColorPickerPopover } from '../ColorPickerPopover.js'
-import { StyleToggleGroup } from '../StyleToggleGroup.js'
+import { CellTypographyFields, CellBorderPaddingFields } from './TableCellStyleFields.js'
 
 interface Props {
   field: TableField
@@ -11,10 +10,9 @@ interface Props {
 }
 
 /**
- * Header / Row / Cell style sub-panels for the table field properties.
- * Extracted from `LoopFieldProps.tsx` to keep that file under the 300-line
- * cap (Hard Rule #11). Behaviour is unchanged versus the inline version
- * apart from the GH #76 transparent-fill controls.
+ * Header / Row / Border / Cell style sub-panels for the table field
+ * properties. The shared font/colour/border field groups live in
+ * `TableCellStyleFields` (Hard Rule #11 + reuse).
  */
 export function TableStyleSections({ field, allFontFamilies }: Props) {
   const updateFieldStyle = useTemplateStore((s) => s.updateFieldStyle)
@@ -38,45 +36,14 @@ export function TableStyleSections({ field, allFontFamilies }: Props) {
     <>
       <div className="tg-panel-section">
         <div className="tg-panel-section-title">Header Style</div>
-
-        <div className="tg-form-row">
-          <label>Font Family</label>
-          <select
-            className="tg-select"
-            value={headerStyle.fontFamily}
-            onChange={(e) => updateHeader({ fontFamily: e.target.value })}
-          >
-            {allFontFamilies.map((f) => (
-              <option key={f} value={f}>
-                {f}
-              </option>
-            ))}
-          </select>
-        </div>
-
-        <div className="tg-form-row">
-          <label>Font Size</label>
-          <NumberInput
-            min={1}
-            value={headerStyle.fontSize}
-            defaultValue={10}
-            onChange={(v) => updateHeader({ fontSize: v })}
-          />
-        </div>
-
-        <div className="tg-form-row">
-          <label>Style</label>
-          <StyleToggleGroup
-            value={{
-              fontWeight: headerStyle.fontWeight,
-              fontStyle: headerStyle.fontStyle,
-              textDecoration: headerStyle.textDecoration,
-            }}
-            onChange={(patch) => updateHeader(patch)}
-            testIdPrefix="table-header"
-          />
-        </div>
-
+        <CellTypographyFields
+          style={headerStyle}
+          allFontFamilies={allFontFamilies}
+          onChange={updateHeader}
+          testIdPrefix="table-header"
+          textColorLabel="Header Text Color"
+          bgLabel="Header Row Background"
+        />
         <div className="tg-form-row">
           <label>Align</label>
           <select
@@ -89,85 +56,18 @@ export function TableStyleSections({ field, allFontFamilies }: Props) {
             <option value="right">Right</option>
           </select>
         </div>
-
-        <div className="tg-form-row">
-          <label>Header Text Color</label>
-          <ColorPickerPopover
-            value={headerStyle.color}
-            onChange={(c) => updateHeader({ color: c })}
-            ariaLabel="Header text color"
-          />
-        </div>
-
-        <div className="tg-form-row">
-          <label>Header Row Background</label>
-          <NullableColorInput
-            value={headerStyle.backgroundColor}
-            onChange={(v) => updateHeader({ backgroundColor: v })}
-            ariaLabel="Header row background color"
-          />
-        </div>
       </div>
 
       <div className="tg-panel-section">
         <div className="tg-panel-section-title">Row Style</div>
-
-        <div className="tg-form-row">
-          <label>Font Family</label>
-          <select
-            className="tg-select"
-            value={rowStyle.fontFamily}
-            onChange={(e) => updateRow({ fontFamily: e.target.value })}
-          >
-            {allFontFamilies.map((f) => (
-              <option key={f} value={f}>
-                {f}
-              </option>
-            ))}
-          </select>
-        </div>
-
-        <div className="tg-form-row">
-          <label>Font Size</label>
-          <NumberInput
-            min={1}
-            value={rowStyle.fontSize}
-            defaultValue={10}
-            onChange={(v) => updateRow({ fontSize: v })}
-          />
-        </div>
-
-        <div className="tg-form-row">
-          <label>Style</label>
-          <StyleToggleGroup
-            value={{
-              fontWeight: rowStyle.fontWeight,
-              fontStyle: rowStyle.fontStyle,
-              textDecoration: rowStyle.textDecoration,
-            }}
-            onChange={(patch) => updateRow(patch)}
-            testIdPrefix="table-row"
-          />
-        </div>
-
-        <div className="tg-form-row">
-          <label>Row Text Color</label>
-          <ColorPickerPopover
-            value={rowStyle.color}
-            onChange={(c) => updateRow({ color: c })}
-            ariaLabel="Row text color"
-          />
-        </div>
-
-        <div className="tg-form-row">
-          <label>Row Background</label>
-          <NullableColorInput
-            value={rowStyle.backgroundColor}
-            onChange={(v) => updateRow({ backgroundColor: v })}
-            ariaLabel="Row background color"
-          />
-        </div>
-
+        <CellTypographyFields
+          style={rowStyle}
+          allFontFamilies={allFontFamilies}
+          onChange={updateRow}
+          testIdPrefix="table-row"
+          textColorLabel="Row Text Color"
+          bgLabel="Row Background"
+        />
         <div className="tg-form-row">
           <label>Overflow Mode</label>
           <select
@@ -214,68 +114,7 @@ export function TableStyleSections({ field, allFontFamilies }: Props) {
 
       <div className="tg-panel-section">
         <div className="tg-panel-section-title">Cell Style</div>
-
-        <div className="tg-form-row">
-          <label>Cell Border Width</label>
-          <NumberInput
-            min={0}
-            step={0.5}
-            value={rowStyle.borderWidth}
-            defaultValue={1}
-            onChange={(v) => updateRow({ borderWidth: v })}
-          />
-        </div>
-
-        <div className="tg-form-row">
-          <label>Cell Border Color</label>
-          <NullableColorInput
-            value={rowStyle.borderColor}
-            onChange={(v) => updateRow({ borderColor: v })}
-            ariaLabel="Cell border color"
-          />
-        </div>
-
-        <div style={{ display: 'grid', gridTemplateColumns: '1fr 1fr', gap: 8 }}>
-          <div className="tg-form-row">
-            <label>Padding Top</label>
-            <NumberInput
-              min={0}
-              value={rowStyle.paddingTop}
-              defaultValue={4}
-              onChange={(v) => updateRow({ paddingTop: v })}
-            />
-          </div>
-
-          <div className="tg-form-row">
-            <label>Padding Bottom</label>
-            <NumberInput
-              min={0}
-              value={rowStyle.paddingBottom}
-              defaultValue={4}
-              onChange={(v) => updateRow({ paddingBottom: v })}
-            />
-          </div>
-
-          <div className="tg-form-row">
-            <label>Padding Left</label>
-            <NumberInput
-              min={0}
-              value={rowStyle.paddingLeft}
-              defaultValue={4}
-              onChange={(v) => updateRow({ paddingLeft: v })}
-            />
-          </div>
-
-          <div className="tg-form-row">
-            <label>Padding Right</label>
-            <NumberInput
-              min={0}
-              value={rowStyle.paddingRight}
-              defaultValue={4}
-              onChange={(v) => updateRow({ paddingRight: v })}
-            />
-          </div>
-        </div>
+        <CellBorderPaddingFields style={rowStyle} onChange={updateRow} />
       </div>
     </>
   )
